Wait for player confirmation before anteing
This commit introduces a mechanic that has players
shake their devices in order to ante before proceeding
with the game.

It has one known problem to be fixed in upcoming commits.
When a player spins the dreidel and lands on something
that produces an ante request (i.e. gimmel), they are anted
almost immediately. This is caused by continued momentum
in the accelerometer from the first shake, and will be
fixed ASAP.